Release first functional version of spymaster game

+++ src/constants.rs
@@ -1,3 +1,3 @@
 pub const WORD_LIST_FILE: &str = "./data/wordlist-eng.txt";
 pub const BOARD_SIZE: usize = 5;
-pub const NUM_WORDS_TO_GUESS: usize = 9;
+pub const NUM_WORDS_TO_GUESS: i8 = 9;
